+++ Project/STM32L1xx_StdPeriph_Templates/Driver/system/delay/delay.h
@@ -2,7 +2,7 @@
 #define __DELAY_H 			   
 #include "stm32l1xx.h" 
 void delay_init(uint8_t SYSCLK);
-void delay_ms(uint16_t nms);
+void delay_ms(uint32_t nms);
 void delay_us(uint32_t nus);
 
 #endif
